remove annotation test controllers
